backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
praw==7.7.1
httpx==0.27.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.0
